requirements/base.txt
Django==3.0.2
djangorestframework==3.11.0
celery==4.4.0
psycopg2-binary==2.8.4
requests==2.22.0
lxml==4.4.2
django-celery-beat==1.5.0
